+++ src/command/owner-commands/suggest.ts
@@ -5,8 +5,8 @@ const command: Command = {
   ownerOnly: true,
   platforms: [Platform.Twitch],
   run: async (ctx, { commandState, services }) => {
-    if (!commandState.wouldBeCommands[ctx.command]) {
-      commandState.wouldBeCommands[ctx.command] =
+    if (!commandState.wouldBeCommands[ctx.body]) {
+      commandState.wouldBeCommands[ctx.body] =
         await services.openaiService.sendPrompt(
           `Tell the user that "${ctx.body}" isn't a command but tell them what it might be if it were`,
           {
@@ -15,7 +15,7 @@ const command: Command = {
           }
         );
     }
-    ctx.botSpeak(' ' + commandState.wouldBeCommands[ctx.command]);
+    ctx.botSpeak(' ' + commandState.wouldBeCommands[ctx.body]);
     return true;
   }
 };